backend/requirements.txt

Flask==3.0.3
mysql-connector-python==8.3.0

